Include event coordinates in callback arguments
For the most part, developers will almost always be wanting to check the
coordinates of an event in a callback. Multiple touches are accommodated
by storing their coordinates in additional indices:

    [ x1, y1, x2, y2, x3, y3, y4, y4 ... ]

This works much better than forcing an author to constantly call utility
methods in every callback to retrieve an event's coordinates.