Use Alhadis/Fix-IE@7b90e8a for bundled polyfills
A dedicated repo is now being used to hold and maintain these polyfills,
instead of keeping them sitting in a snippet dumping ground. This commit
was brought into being as a result of changing hardlinks on my system to
point to the new repo. Nothing about the logic has otherwise changed.