Awkwardly attempt to modernise project with new APIs
The Pointer Events specification hadn't even left draft status when this
library was first written. In a sense, PointerEvents solved the problems
that Gestures was designed to do.

Might be so much easier and cleaner to eradicate the original `Gestures`
and write the base component layer from scratch. We should see if that's
even needed first, though.